backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic==2.10.0
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
httpx==0.27.2
pytest==8.3.3
fpdf2==2.8.2
qrcode[pil]==8.0
